refactor(painting): group searchBar logic
