+++ src/app/painting/page.tsx
@@ -1,10 +1,10 @@
 import React, { Suspense } from "react";
-import { PaintingCardGrid } from "../../components/painting/PaintingCardGrid";
+import { PaintingCardGrid } from "../../components/painting/cardGrid/PaintingCardGrid";
 import { findPaintingAction } from "../../server-action/backend/painting/api";
 import { ErrorModal } from "../../components/modal/ErrorModal";
-import { getSearchParams } from "../../components/painting/util";
-import { PAINTING_PARAM_KEY } from "../../components/painting/const";
-import { PaintingSearchBar } from "../../components/painting/PaintingSearchbar";
+import { getSearchParams } from "../../components/painting/searchBar/util";
+import { PAINTING_PARAM_KEY } from "../../components/painting/searchBar/const";
+import { PaintingSearchBar } from "../../components/painting/searchBar/PaintingSearchbar";
 
 // TODO: Search Page 개선
 // - [x] '/'page에 search bar 추가
