+++ src/components/painting/searchBar/const.ts
