+++ src/components/painting/searchBar/type.ts
